+++ NEWS.md
@@ -2,6 +2,7 @@
 * Added `as.dendrogram.hdbscan` function to make hdbscan objects compatible with other R hierarchical clustering implementations. 
 * New utility function `sgdBatches` helps estimate training time for datasets. 
 * Fixed bug in estimation of `sgd_batches` where 10x to many batches would be used for dataset < 10000 nodes. 
+* `buildEdgeMatrix` and `distance` now store the `distance_method` in attribute `method` of the returned object. 
 
 ### largeVis 0.1.9.2
 Hotfix for a bug in the neighbor search when max iterations was 0.